requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-phonenumber-field==7.0.2
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.6
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
